collector/requirements.txt
omron-2jcie-bu01
influxdb-client
python-dotenv
requests
